Correção de bugs e adição de dependências

+++ lib/models/section.dart
@@ -95,11 +95,11 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems) {
-      if (!items.contains(original)) {
+      if (!items.contains(original) && (original.image as String).contains('firebase')) {
         try {
           final ref = storage.refFromURL(original.image as String);
           await ref.delete();
-        } catch (error) {}
+        } catch (e) {}
       }
     }
 
@@ -113,10 +113,12 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = storage.refFromURL(item.image as String);
-        await ref.delete();
-      } catch (error) {}
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = storage.refFromURL(item.image as String);
+          await ref.delete();
+        } catch (e) {}
+      }
     }
   }
